Split Services bindings into smaller compile units.

+++ src/components/ogre/widgets/EntityIconManager.cpp
@@ -50,7 +50,7 @@ EntityIconSlot* EntityIconManager::createSlot(unsigned int pixelSize) {
 	ss << "entityIconSlot" << mSlotsCounter++;
 	//Make the slot more visible.
 	UniqueWindowPtr<CEGUI::Window> container(mGuiManager.createWindow("EmberLook/StaticImage", ss.str()));
-	container->setSize(CEGUI::USize(CEGUI::UDim(0, pixelSize), CEGUI::UDim(0, pixelSize)));
+	container->setSize(CEGUI::USize(CEGUI::UDim(0, (float)pixelSize), CEGUI::UDim(0, (float)pixelSize)));
 	auto slot = std::make_unique<EntityIconSlot>(std::move(container));
 	mSlots.emplace_back(std::move(slot));
 	return mSlots.back().get();
@@ -71,7 +71,7 @@ std::unique_ptr<EntityIcon> EntityIconManager::createIconInstance(Gui::Icons::Ic
 	UniqueWindowPtr<CEGUI::Window> amountWindow(mGuiManager.createWindow("EmberLook/StaticText", ss.str()));
 
 	if (item && iconWindow && amountWindow) {
-		item->setSize(CEGUI::USize(CEGUI::UDim(0, pixelSize), CEGUI::UDim(0, pixelSize)));
+		item->setSize(CEGUI::USize(CEGUI::UDim(0, (float)pixelSize), CEGUI::UDim(0, (float)pixelSize)));
 		//item->setTooltipText(name);
 
 		iconWindow->setProperty("BackgroundEnabled", "false");
